requirements.txt
openai==0.27.5
python-dotenv==1.0.0
